# src/main/resources/openAPI_v002.yaml
openapi: 3.1.0
info:
  title: Agricultural federation API
  description: |
    # Versioning
    This API uses semantic versioning (MAJOR.MINOR.PATCH). Available versions are :
    - v0.0.1 (2026-04-21) : POST /collectivities for collectivities creation and POST /members for members creation.
    - v0.0.2 (2026-04-22) : PATCH /collectivities/{id} for assigning a unique number and name to a collectivity.
  version: 0.0.2
servers:
  - url: 'https://localhost:8080'
paths:
  '/collectivities':
    post:
      summary: Create list of collectivities
      description: |
        During collectivity request body creation, only members ID are provided to identify members.
        But the response content includes members information, not only their IDs.
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateCollectivity'
      responses:
        201:
          description: List of created collectivities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Collectivity'
        400:
          description: Collectivity without federation approval or structure missing.
        404:
          description: Member not found.

  '/collectivities/{id}':
    patch:
      summary: Assign a unique number and name to a collectivity
      description: |
        This endpoint allows the federation to assign a unique number and a unique name to
        an existing collectivity. Once a number or a name has been assigned, it can no longer
        be changed. Attempting to update an already assigned number or name will result in a
        409 Conflict response.
        The system ensures that the provided name does not already exist among other collectivities.
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the collectivity to update.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignCollectivityIdentity'
      responses:
        200:
          description: Collectivity successfully updated with its number and name.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collectivity'
        400:
          description: Request body is missing required fields (number or name).
        404:
          description: Collectivity not found.
        409:
          description: |
            Conflict. One of the following cases:
            - The collectivity already has a number assigned and cannot be changed.
            - The collectivity already has a name assigned and cannot be changed.
            - The provided name already exists in another collectivity.

  '/members':
    post:
      summary: Create list of members
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateMember'
      responses:
        201:
          description: List of created members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
        400:
          description: |
            The following case must return a bad request exception :
            - Member with bad referees or without proper payment.
            - Membership dues not paid or registration fee not paid.
        404:
          description: Either collectivity or member not found.

components:
  schemas:
    CreateCollectivity:
      type: object
      properties:
        location:
          type: string
        members:
          type: array
          items:
            schema:
              $ref: '#/components/schemas/MemberIdentifier'
        federationApproval:
          type: boolean
        structure:
          $ref: '#/components/schemas/CreateCollectivityStructure'

    AssignCollectivityIdentity:
      type: object
      required:
        - number
        - name
      properties:
        number:
          type: string
          description: |
            Unique number assigned by the federation to identify the collectivity.
            Once assigned, it cannot be changed.
          example: "COL-0042"
        name:
          type: string
          description: |
            Unique name assigned by the federation to identify the collectivity.
            Once assigned, it cannot be changed. The system verifies uniqueness.
          example: "Collectivité Betsileo Nord"

    CreateCollectivityStructure:
      type: object
      properties:
        president:
          $ref: '#/components/schemas/MemberIdentifier'
        vicePresident:
          $ref: '#/components/schemas/MemberIdentifier'
        treasurer:
          $ref: '#/components/schemas/MemberIdentifier'
        secretary:
          $ref: '#/components/schemas/MemberIdentifier'

    CollectivityStructure:
      type: object
      properties:
        president:
          $ref: '#/components/schemas/Member'
        vicePresident:
          $ref: '#/components/schemas/Member'
        treasurer:
          $ref: '#/components/schemas/Member'
        secretary:
          $ref: '#/components/schemas/Member'

    Collectivity:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
          description: Unique number assigned by the federation. Null if not yet assigned.
        name:
          type: string
          description: Unique name assigned by the federation. Null if not yet assigned.
        location:
          type: string
        structure:
          $ref: '#/components/schemas/CollectivityStructure'
        members:
          type: array
          items:
            $ref: '#/components/schemas/Member'

    MemberIdentifier:
      type: string

    CreateMember:
      type: object
      allOf:
        - $ref: '#/components/schemas/MemberInformation'
      properties:
        collectivityIdentifier:
          type: string
        referees:
          type: array
          items:
            $ref: '#/components/schemas/MemberIdentifier'
        registrationFeePaid:
          type: boolean
        membershipDuesPaid:
          type: boolean

    MemberInformation:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          $ref: '#/components/schemas/Gender'
        address:
          type: string
        profession:
          type: string
        phoneNumber:
          type: integer
        email:
          type: string
        occupation:
          $ref: '#/components/schemas/MemberOccupation'

    Member:
      type: object
      allOf:
        - $ref: '#/components/schemas/MemberInformation'
      properties:
        id:
          $ref: '#/components/schemas/MemberIdentifier'
        referees:
          type: array
          items:
            $ref: '#/components/schemas/Member'

    Gender:
      type: string
      enum:
        - MALE
        - FEMALE

    MemberOccupation:
      type: string
      enum:
        - JUNIOR
        - SENIOR
        - SECRETARY
        - TREASURER
        - VICE_PRESIDENT
        - PRESIDENT